requirements.txt
discord.py[voice]
pydub
openai
whisper
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
